+++ plugins/postcss-random-function/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes to PostCSS Random Function
 
+### Unreleased (major)
+
+- Updated: Support for Node `20.19.0` or later (major).
+- Removed: `commonjs` API. In supported Node versions `require(esm)` will work without needing to make code changes.
+
 ### 2.0.1
 
 _May 27, 2025_